Allow test explorer to resolve all tests under a path in nested mode
Summary:
adding a feature that allows a user to select a directory and resolve all test within the selected directory

When "packageDisplayMode" is nested, tests will now nest properly under
each other (previously not the case as go.mod may not be at the Bazel
workspace root.  Regardless of package mode, user will see a new right
click option on a test case to index all tests under that path.